+++ discopy/traced.py
@@ -134,8 +134,6 @@ from discopy.utils import (
     assert_isinstance,
     assert_istraceable,
 )
-from discopy.testing import inapplicable
-
 FREE_TRACE = "A free trace is a box, not a rewrite."
 
 
@@ -176,17 +174,23 @@ class Diagram(monoidal.Diagram, TracedCategory):
     def to_drawing(self):
         return monoidal.Diagram.to_drawing(self, functor_factory=Functor)
 
-    trace_dinaturality_left = inapplicable(FREE_TRACE)
+    trace_dinaturality_left = \
+        TracedCategory.trace_dinaturality_left.inapplicable(FREE_TRACE)
 
-    trace_dinaturality_right = inapplicable(FREE_TRACE)
+    trace_dinaturality_right = \
+        TracedCategory.trace_dinaturality_right.inapplicable(FREE_TRACE)
 
-    trace_naturality_left = inapplicable(FREE_TRACE)
+    trace_naturality_left = \
+        TracedCategory.trace_naturality_left.inapplicable(FREE_TRACE)
 
-    trace_naturality_right = inapplicable(FREE_TRACE)
+    trace_naturality_right = \
+        TracedCategory.trace_naturality_right.inapplicable(FREE_TRACE)
 
-    trace_superposing_left = inapplicable(FREE_TRACE)
+    trace_superposing_left = \
+        TracedCategory.trace_superposing_left.inapplicable(FREE_TRACE)
 
-    trace_superposing_right = inapplicable(FREE_TRACE)
+    trace_superposing_right = \
+        TracedCategory.trace_superposing_right.inapplicable(FREE_TRACE)
 
 
 class Box(monoidal.Box, Diagram):